+++ src/api/run_cucumber_spec.ts
@@ -1,44 +1,10 @@
-import { Envelope, TestStepResultStatus, IdGenerator } from '@cucumber/messages'
-import fs from 'mz/fs'
-import path from 'path'
-import { reindent } from 'reindent-template-literals'
-import { PassThrough } from 'stream'
+import { Envelope, TestStepResultStatus } from '@cucumber/messages'
 import { expect } from 'chai'
 import { runCucumber } from './run_cucumber'
 import { IRunEnvironment } from './types'
 import { loadSupport } from './load_support'
 import { loadConfiguration } from './load_configuration'
-
-const newId = IdGenerator.uuid()
-
-async function setupEnvironment(): Promise<Partial<IRunEnvironment>> {
-  const cwd = path.join(__dirname, '..', '..', 'tmp', `runCucumber_${newId()}`)
-  await fs.mkdir(path.join(cwd, 'features'), { recursive: true })
-  await fs.writeFile(
-    path.join(cwd, 'features', 'test.feature'),
-    reindent(`Feature: test fixture
-      Scenario: one
-        Given a step
-        Then another step`)
-  )
-  await fs.writeFile(
-    path.join(cwd, 'features', 'steps.ts'),
-    reindent(`import { Given, Then } from '../../../src'
-    Given('a step', function () {})
-    Then('another step', function () {})`)
-  )
-  await fs.writeFile(
-    path.join(cwd, 'cucumber.mjs'),
-    `export default {paths: ['features/test.feature'], requireModule: ['ts-node/register'], require: ['features/steps.ts']}`
-  )
-  const stdout = new PassThrough()
-  return { cwd, stdout }
-}
-
-async function teardownEnvironment(environment: IRunEnvironment) {
-  await fs.rmdir(environment.cwd, { recursive: true })
-  environment.stdout.end()
-}
+import { setupEnvironment, teardownEnvironment } from './test_helpers'
 
 describe('runCucumber', () => {
   describe('preloading support code', () => {
