+++ src/api/test_helpers.ts
@@ -1,0 +1,37 @@
+import { IRunEnvironment } from "./types";
+import path from "path";
+import fs from "mz/fs";
+import { reindent } from "reindent-template-literals";
+import { PassThrough } from "stream";
+import { IdGenerator } from "@cucumber/messages";
+
+const newId = IdGenerator.uuid()
+
+export async function setupEnvironment(): Promise<Partial<IRunEnvironment>> {
+  const cwd = path.join(__dirname, '..', '..', 'tmp', `api_${newId()}`)
+  await fs.mkdir(path.join(cwd, 'features'), { recursive: true })
+  await fs.writeFile(
+    path.join(cwd, 'features', 'test.feature'),
+    reindent(`Feature: test fixture
+      Scenario: one
+        Given a step
+        Then another step`)
+  )
+  await fs.writeFile(
+    path.join(cwd, 'features', 'steps.ts'),
+    reindent(`import { Given, Then } from '../../../src'
+    Given('a step', function () {})
+    Then('another step', function () {})`)
+  )
+  await fs.writeFile(
+    path.join(cwd, 'cucumber.mjs'),
+    `export default {paths: ['features/test.feature'], requireModule: ['ts-node/register'], require: ['features/steps.ts']}`
+  )
+  const stdout = new PassThrough()
+  return { cwd, stdout }
+}
+
+export async function teardownEnvironment(environment: IRunEnvironment) {
+  await fs.rmdir(environment.cwd, { recursive: true })
+  environment.stdout.end()
+}
